// src/app/enum/actions-type.enum.ts
export enum BUTTON_ACTION_ENUM { 
    EDIT = 'edit',
    CREATE = 'create',
    DELETE = 'delete',
    VIEW = 'view'
}

export enum BUTTON_CLASS_ENUM {
    EDIT = 'btn btn-primary',
    CREATE = 'create',
    DELETE = 'btn btn-danger',
    VIEW = 'view'
}

export enum BUTTON_ICON_ENUM {
    EDIT = 'bi bi-pencil-square text-primary',
    CREATE = 'create',
    DELETE = 'bi bi-trash text-danger',
    VIEW = 'view'
}